pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "campaignpilot"
version = "0.1.0"
description = "Infrastructure layer for CampaignPilot"
requires-python = ">=3.11"

dependencies = [
    "anthropic>=0.40.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "chromadb>=0.5.0",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "streamlit>=1.40.0",
    "faker>=33.0.0",
    "python-dotenv>=1.0.1",
    "langfuse>=2.50.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "rich>=13.9.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "numpy>=2.1.0",
    "scipy>=1.14.0",
    "pandas>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "ruff",
    "mypy",
]
